requirements.txt
apify-client==0.6.0
azure-core==1.26.1
azure-functions==1.12.0
azure-storage-blob==12.14.1
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
idna==3.4
isodate==0.6.1
msrest==0.7.1
numpy==1.23.5
oauthlib==3.2.2
pandas==1.5.2
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.12
